requirements.txt
numpy==1.24.2
opencv_python==4.6.0.66
pyserial==3.5
